Improvements from PR feedback and device testing
- Change x-threshold for swipe motion from 150 to 50px
- Move swipe detection code to its own file
- Allow 'prev' parameter for showDictionary function
- Add @stylistic/eslint-plugin-js to handle max-len linting for comments
